+++ include/roq/portfolio_update.hpp
@@ -24,9 +24,9 @@ namespace roq {
 
 //! Portfolio update  !!! EXPERIMENTAL !!!
 struct ROQ_PUBLIC PortfolioUpdate final {
-  std::string_view account;                         //!< Account name
   std::string_view user;                            //!< User (optional)
-  uint32_t strategy = {};                           //!< Strategy (optional)
+  uint32_t strategy_id = {};                        //!< Strategy identifier (optional)
+  std::string_view account;                         //!< Account name
   std::span<Position const> positions;              //!< Position updates
   UpdateType update_type = {};                      //!< Update type
   std::chrono::nanoseconds exchange_time_utc = {};  //!< Exchange timestamp, possibly from matching engine (UTC)
@@ -53,16 +53,16 @@ struct fmt::formatter<roq::PortfolioUpdate> {
     return fmt::format_to(
         context.out(),
         R"({{)"
-        R"(account="{}", )"
         R"(user="{}", )"
-        R"(strategy={}, )"
+        R"(strategy_id={}, )"
+        R"(account="{}", )"
         R"(positions=[{}], )"
         R"(update_type={}, )"
         R"(exchange_time_utc={})"
         R"(}})"_cf,
-        value.account,
         value.user,
-        value.strategy,
+        value.strategy_id,
+        value.account,
         fmt::join(value.positions, ", "sv),
         value.update_type,
         value.exchange_time_utc);