+++ include/roq/trade_update.hpp
@@ -42,6 +42,7 @@ struct ROQ_PUBLIC TradeUpdate final {
   UpdateType update_type = {};                     //!< Update type
   std::chrono::nanoseconds sending_time_utc = {};  //!< Exchange sending timestamp (UTC)
   std::string_view user;                           //!< User name (optional, only relevant for drop-copy)
+  uint32_t strategy_id = {};                       //!< Strategy identifier (optional)
 };
 
 template <>
@@ -80,7 +81,8 @@ struct fmt::formatter<roq::TradeUpdate> {
         R"(routing_id="{}", )"
         R"(update_type={}, )"
         R"(sending_time_utc={}, )"
-        R"(user="{}")"
+        R"(user="{}", )"
+        R"(strategy_id={})"
         R"(}})"_cf,
         value.stream_id,
         value.account,
@@ -97,7 +99,8 @@ struct fmt::formatter<roq::TradeUpdate> {
         value.routing_id,
         value.update_type,
         value.sending_time_utc,
-        value.user);
+        value.user,
+        value.strategy_id);
   }
 };